Bug fix: bring all dependencies as a temporary fix.
