+++ src/bundlers/typescript/nodeTsBundler.ts
@@ -43,85 +43,6 @@ export class NodeTsBundler implements BundlerInterface {
         }
     }
 
-    async #getNodeModulesTs(
-        filePath: string,
-        mode: "development" | "production"
-    ): Promise<any> {
-
-        if (mode === "development") {
-            return null;
-        }
-
-        const dependencies: string[] = [];
-        const { name } = getFileDetails(filePath);
-        const outputFile = `${name}-processed.js`;
-        const temporaryFolder = await createTemporaryFolder();
-
-        const module = {
-            rules: [
-                {
-                    test: /\.tsx?$/,
-                    use: [
-                        {
-                            loader: "ts-loader",
-                            options: {
-                                configFile: "tsconfig.json",
-                                onlyCompileBundledFiles: true
-                            }
-                        }
-                    ],
-                    exclude: /really\.html/
-                }
-            ]
-        };
-
-        const plugins = [
-            new NodePolyfillPlugin(),
-            new AccessDependenciesPlugin(dependencies, process.cwd()),
-        ];
-        const resolve = { extensions: [".tsx", ".ts", ".js"] };
-        const resolveLoader = {
-            modules: [path.resolve(__dirname, "../../../", "node_modules")],
-            conditionNames: ["require"]
-        };
-
-        await bundle(
-            "./" + filePath,
-            mode,
-            [],
-            module,
-            plugins,
-            temporaryFolder,
-            outputFile,
-            resolve,
-            resolveLoader
-        );
-
-        // delete the temporary folder
-        await deleteFolder(temporaryFolder);
-
-        const dependenciesInfo = dependencies.map((dependency) => {
-            const relativePath = dependency.split("node_modules" + path.sep)[1];
-            const dependencyName = relativePath?.split(path.sep)[0];
-            const dependencyPath =
-                dependency.split("node_modules" + path.sep)[0] +
-                "node_modules" +
-                path.sep +
-                dependencyName;
-            //dependencyPath.replace(folder, cwd);
-            return {
-                name: dependencyName,
-                path: dependencyPath
-            };
-        });
-
-        // remove duplicates from dependenciesInfo by name
-        const uniqueDependenciesInfo = dependenciesInfo.filter(
-            (v, i, a) => a.findIndex((t) => t.name === v.name) === i
-        );
-
-        return uniqueDependenciesInfo;
-    }
     async #copyDependencies(dependenciesInfo: any, tempFolderPath: string, mode: "development" | "production") {
         const nodeModulesPath = path.join(tempFolderPath, "node_modules");
 
@@ -277,8 +198,7 @@ export class NodeTsBundler implements BundlerInterface {
 
         debugLogger.debug(`[NodeTSBundler] Get the list of node modules and bundling the javascript code for file ${input.path}.`)
         // 2. Run webpack to get dependenciesInfo and the packed file
-        const [dependenciesInfo, _] = await Promise.all([
-            this.#getNodeModulesTs(input.path, mode),
+        await Promise.all([
             this.#bundleTypescriptCode(
                 input.configuration.path,
                 temporaryFolder,
@@ -292,7 +212,7 @@ export class NodeTsBundler implements BundlerInterface {
         // 2. Copy non js files and node_modules and write index.js file
         await Promise.all([
             this.#copyNonTsFiles(temporaryFolder),
-            mode === "production" ? this.#copyDependencies(dependenciesInfo, temporaryFolder, mode) : Promise.resolve(),
+            mode === "production" ? this.#copyDependencies(input.extra!.dependenciesInfo, temporaryFolder, mode) : Promise.resolve(),
             writeToFile(temporaryFolder, "index.js", lambdaHandler(`"${input.configuration.name}"`))
         ]);
 
@@ -303,7 +223,7 @@ export class NodeTsBundler implements BundlerInterface {
             ...input,
             path: temporaryFolder,
             extra: {
-                dependenciesInfo
+                dependenciesInfo: input.extra?.dependenciesInfo
             }
         };
     }
