+++ src/models/projectConfiguration.ts
@@ -69,6 +69,7 @@ export class ProjectConfiguration {
     name: string;
     region: string;
     sdk: SdkConfiguration;
+    options?: any;
     cloudProvider: CloudProviderIdentifier;
     astSummary: AstSummary;
     classes: ClassConfiguration[];
@@ -80,6 +81,7 @@ export class ProjectConfiguration {
         this.name = yamlConfiguration.name;
         this.region = yamlConfiguration.region;
         this.sdk = yamlConfiguration.sdk;
+        this.options = yamlConfiguration.options;
         this.cloudProvider = yamlConfiguration.cloudProvider || CloudProviderIdentifier.GENEZIO;
 
         // Generate AST Summary
